Dungeons, spatials, slap crafting refactoring, and hybrid movement adjustments (#353)
first batch of PRs to help with randomizing dungeons, some optimizations, and others.

this includes
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/3017
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/3077
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2294
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/3217
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2469
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2410

and parts from
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2452

This does not pull in the Mathios dungeon files in this PR, as they need to be updated and have the mob count reduced to keep the server stable. This is only to test the start of that effort.